requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
aiogram==3.4.1
pydantic==2.6.4
pydantic-settings==2.2.1
SQLAlchemy==2.0.28
alembic==1.13.1
asyncpg==0.29.0
httpx==0.27.0
beautifulsoup4==4.12.3
readability-lxml==0.8.1
loguru==0.7.2
python-dotenv==1.0.1
